requirements.txt
gdown==4.4.0
pandas==1.4.2
sklearn==0.0
scikit-learn==1.1.0
numpy==1.22.3
PyYAML==6.0
click==8.1.3
mlflow==1.25.1
pytest==7.1.2
py==1.11.0
pip==21.1.2
wheel==0.36.2
setuptools==57.0.0
marshmallow_dataclass==8.5.8